feat: merge config and handle
