+++ nobodywho/core/src/chat/api/mod.rs
@@ -1,5 +1,3 @@
-pub mod config;
 pub mod handle;
 
-pub use config::{ChatBuilder, ChatConfig};
-pub use handle::{ChatHandle, ChatHandleAsync};
+pub use handle::{ChatBuilder, ChatConfig, ChatHandle, ChatHandleAsync};
